simplifying + removing redundant code+ class rename
